base package set to org.cbioportal.cam
