+++ src/main/java/org/cbioportal/cam/ClinicalAttributes.java
@@ -13,19 +13,19 @@
  * Center has been advised of the possibility of such damage.
  */
 
-package org.mskcc.clinical_attributes.service.exception;
+package org.cbioportal.cam;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-public class ClinicalAttributeNotFoundException extends RuntimeException {
+/**
+ * @author Manda Wilson 
+ */
+@SpringBootApplication // shorthand for @Configuration, @EnableAutoConfiguration, @ComponentScan
+public class ClinicalAttributes {
 
-    private static final Logger logger = LoggerFactory.getLogger(ClinicalAttributeNotFoundException.class);
-
-    public ClinicalAttributeNotFoundException(String normalizedColumnHeader) {
-        super("Could not find clinical attribute by normalized column header '" + normalizedColumnHeader + "'.");
-        logger.error("Could not find clinical attribute by normalized column header '" + normalizedColumnHeader + "'.");
+    public static void main(String[] args) {
+        SpringApplication.run(ClinicalAttributes.class, args);
     }
-
 }
