+++ src/main/java/org/cbioportal/cam/service/ClinicalAttributesService.java
@@ -13,15 +13,18 @@
  * Center has been advised of the possibility of such damage.
  */
 
-package org.mskcc.clinical_attributes.repository;
+package org.cbioportal.cam.service;
 
 import java.util.List;
 
-import org.mskcc.clinical_attributes.model.ClinicalAttribute;
+import org.cbioportal.cam.model.ClinicalAttribute;
+import org.cbioportal.cam.service.exception.ClinicalAttributeNotFoundException;
 
 /**
  * @author Manda Wilson 
  */
-public interface ClinicalAttributesRepository {
-    List<ClinicalAttribute> getClinicalAttribute();
+public interface ClinicalAttributesService {
+    List<ClinicalAttribute> getClinicalAttributes(String studyId);
+    List<ClinicalAttribute> getMetadataByNormalizedColumnHeaders(String studyId, List<String> normalizedColumnHeaders) throws ClinicalAttributeNotFoundException;
+    ClinicalAttribute getMetadataByNormalizedColumnHeader(String studyId, String normalizedColumnHeader) throws ClinicalAttributeNotFoundException;
 }
